pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "geo-city"
version = "0.1.0"
description = "Find the nearest city to any geographic coordinates or from photo metadata"
readme = "README.md"
authors = [
    {name = "GeoCity Team", email = "example@geocity.org"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9", 
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Utilities",
    "Topic :: Multimedia :: Graphics",
    "Topic :: Scientific/Engineering :: GIS",
]
requires-python = ">=3.7"
dependencies = [
    "requests>=2.25.0",
    "pillow>=8.0.0",
    "pandas>=1.0.0", 
    "numpy>=1.18.0",
    "appdirs>=1.4.4",
]

[project.urls]
Homepage = "https://github.com/yourusername/geo-city"
Documentation = "https://geocity.readthedocs.io/"
Repository = "https://github.com/yourusername/geo-city.git"
"Bug Tracker" = "https://github.com/yourusername/geo-city/issues"

[project.optional-dependencies]
web = [
    "streamlit>=1.0.0",
    "folium>=0.12.0",
    "streamlit-folium>=0.6.0",
]
dev = [
    "pytest>=6.0.0",
    "black>=22.0.0",
    "flake8>=4.0.0",
    "mypy>=0.900",
    "sphinx>=4.0.0",
]

[project.scripts]
geocity = "geo_city.nearest_city:main"

[tool.black]
line-length = 88
target-version = ["py37", "py38", "py39", "py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"

# uv 配置 (以支持更快的包管理)
[tool.uv]
# uv 特定配置
exclude = ["tests"]
requirements = ["requirements.txt"]
resolution = "highest" 